99_参考资料/工控售后 RAG 知识库/codes/POC/requirements.txt
streamlit==1.42.2
langchain==0.3.18
langchain-core==0.3.36
langchain-community==0.3.17
langchain-openai==0.3.6
langchain-chroma==0.2.2
chromadb==0.6.3
python-dotenv==1.0.1
watchdog==6.0.0
minio==7.2.3
supabase==2.10.0
pytest==8.3.4
